Address Explore PR review findings
- Inspect the durable Explore generation before process caches in
  cache-status, so warm caches never hide a corrupt or stale generation
- Stamp Explore jobs with kind="explore" and reject cross-kind job ids
  in Explore status/cancel, mirroring the pivot side
- Gate Chart Configure freshness and filter-member display on the
  node's current Explore cache identity hash, shared from NodePanel
- Key the member loader closure on that hash so display-only pivot
  edits neither reload members nor race a context ref update
- Remove the no-op include_subtotals chart option end to end
- Prove pivot supersession family scoping with a registry unit test
  and a route test that fails if the family key drops the pivot id
- Extract explorePivots/placements, FilterMemberPicker, and ZoneSection
  from ExplorePivotsConfig and reuse the shared INPUT_STYLE
